working, but needs alot more efficiency
